Motor Control (Teensy 3.1 in Arduino)
The notes txt file details issues and concerns based around the motor
testing from Friday the 28th March. There was a delay in setting up GIT
that has now been resolved, and Xavier's latest response on motor
decisions is that a large enough gear CAN be made to accommodate the
original design.